+++ src/chacha/node.ts
@@ -1,7 +1,7 @@
 import { Cipher, u32, u8 } from "@noble/ciphers/utils";
 
 import { _compat } from "../_node/compat";
-import { _hchacha } from "../_node/hchacha";
+import { _hchacha20 } from "../_node/hchacha";
 
 export const xchacha20 = (
   key: Uint8Array,
@@ -14,10 +14,21 @@ export const xchacha20 = (
   const constants = new Uint32Array([0x61707865, 0x3320646e, 0x79622d32, 0x6b206574]); // "expand 32-byte k"
   const subKey = new Uint32Array(8);
 
-  _hchacha(constants, u32(key), u32(nonce.subarray(0, 16)), subKey);
+  _hchacha20(constants, u32(key), u32(nonce.subarray(0, 16)), subKey);
 
   const subNonce = new Uint8Array(12);
   subNonce.set([0, 0, 0, 0]);
   subNonce.set(nonce.subarray(16), 4);
   return _compat("chacha20-poly1305", u8(subKey), subNonce, AAD);
+};
+
+export const chacha20 = (
+  key: Uint8Array,
+  nonce: Uint8Array,
+  AAD?: Uint8Array
+): Cipher => {
+  if (nonce.length !== 12) {
+    throw new Error("chacha20's nonce must be 12 bytes");
+  }
+  return _compat("chacha20-poly1305", key, nonce, AAD);
 };