+++ CHANGELOG.md
@@ -1,13 +1,19 @@
 # Release Notes
 
+## 0.2.2
+
+- Add `chacha20-poly1305` support
+- Bump dependencies
+- Add minimum supported node (16.0.0) runtime check in CI
+
 ## 0.2.1
 
 - Add React Native package exports
 
 ## 0.2.0
 
-- Add xchacha20-poly1305 support
+- Add `xchacha20-poly1305` support
 
 ## 0.1.0
 
-- First beta version release with aes-256-gcm and aes-256-cbc support
+- First beta version release with `aes-256-gcm` and `aes-256-cbc` support